+++ src/autocomplete.c
@@ -6,7 +6,7 @@ char* custom_input(Tree* rules, char* title, COLOR_TYPE title_color, COLOR_TYPE 
     // Initialize buffer for reading
     short buff_len = 0;
     short buff_cap = terminal_width();
-    char* buff = (char*)calloc(buff_cap, sizeof(char));
+    char* buff = (char*)calloc((size_t)buff_cap, sizeof(char));
     if (buff == NULL) {
         fprintf(stderr, "[ERROR] Couldn't allocate memory for buffer\n");
         exit(1);
@@ -86,7 +86,7 @@ char* custom_input(Tree* rules, char* title, COLOR_TYPE title_color, COLOR_TYPE 
         else if (ch == BACKSPACE) {
             if (buff_len != 0 && buff_len - offset >= 1) {
                 // Delete character from buffer
-                for (unsigned i = buff_len - offset - 1; i < buff_cap - 1; i++) {
+                for (unsigned i = (unsigned int)(buff_len - offset - 1); i < buff_cap - 1; i++) {
                     buff[i] = buff[i + 1];
                 }
                 buff_len -= 1;
@@ -97,7 +97,7 @@ char* custom_input(Tree* rules, char* title, COLOR_TYPE title_color, COLOR_TYPE 
         else if (ch == TAB) {
             if (pred->type != FAILURE) {
                 char* prediction = (char*)vector_get(pred->tokens, hint_num % pred->tokens->length);
-                unsigned predict_len = strlen(prediction);
+                unsigned predict_len = (unsigned int)strlen(prediction);
 
                 // Make sure the candidate has no optional brackets
                 if (!contain_chars(prediction, optional_brackets)) {
@@ -115,7 +115,7 @@ char* custom_input(Tree* rules, char* title, COLOR_TYPE title_color, COLOR_TYPE 
                     }
 
                     // Fix buffer length
-                    buff_len = (short)(buff_len - space_offset);
+                    buff_len = buff_len - space_offset;
 
                     // Append end characters to buffer
                     buff[buff_len++] = ' ';
@@ -161,7 +161,7 @@ char* custom_input(Tree* rules, char* title, COLOR_TYPE title_color, COLOR_TYPE 
                     {
                         if (buff_len != 0 && offset != 0) {
                             // Delete character from buffer
-                            for (unsigned i = buff_len - offset; i < buff_cap - 1; i++) {
+                            for (unsigned i = (unsigned int)(buff_len - offset); i < buff_cap - 1; i++) {
                                 buff[i] = buff[i + 1];
                             }
                             buff_len -= 1;
